fix: add null-guards to dialog close callbacks in savings, clients, and loans views to prevent runtime errors (#3551)

+++ src/app/clients/clients-view/clients-view.component.ts
@@ -277,6 +277,9 @@ export class ClientsViewComponent implements OnInit {
       data: { deleteContext: `client with id: ${this.clientViewData.id}` }
     });
     deleteClientDialogRef.afterClosed().subscribe((response: any) => {
+      if (!response) {
+        return;
+      }
       if (response.delete) {
         this.clientsService.deleteClient(this.clientViewData.id).subscribe(() => {
           this.router.navigate(['/clients'], { relativeTo: this.route });
@@ -291,6 +294,9 @@ export class ClientsViewComponent implements OnInit {
   private unassignStaff() {
     const unAssignStaffDialogRef = this.dialog.open(UnassignStaffDialogComponent);
     unAssignStaffDialogRef.afterClosed().subscribe((response: { confirm: any }) => {
+      if (!response) {
+        return;
+      }
       if (response.confirm) {
         this.clientsService
           .executeClientCommand(this.clientViewData.id, 'unassignStaff', { staffId: this.clientViewData.staffId })
@@ -313,6 +319,9 @@ export class ClientsViewComponent implements OnInit {
         }
       });
       viewSignatureDialogRef.afterClosed().subscribe((response: any) => {
+        if (!response) {
+          return;
+        }
         if (response.upload) {
           this.uploadSignature();
         } else if (response.draw) {
@@ -361,6 +370,9 @@ export class ClientsViewComponent implements OnInit {
         data: documents
       });
       deleteSignatureDialogRef.afterClosed().subscribe((response: any) => {
+        if (!response) {
+          return;
+        }
         if (response.delete) {
           this.clientsService.deleteClientDocument(this.clientViewData.id, response.id).subscribe(() => {
             this.reload();
